Force name change.
