+++ Hestia/Hestia/Frontenda/Auth0/ServerSelectScreen/UIViewControllerEditServerName.cs
