+++ Hestia/Hestia/Frontenda/Auth0/ServerSelectScreen/ViewControllerServerList.cs
