+++ Hestia/Hestia/Frontenda/DevicesScreen/ActivatorScreen/PopoverDelegate.cs
