+++ Hestia/Hestia/Frontenda/DevicesScreen/ActivatorScreen/UITableViewActivators.cs
