+++ Hestia/Hestia/Frontenda/DevicesScreen/AddDeviceScreen/PropertyCell.cs
