+++ Hestia/Hestia/Frontenda/DevicesScreen/AddDeviceScreen/TableSourceAddDeviceProperties.cs
