+++ Hestia/Hestia/Frontenda/DevicesScreen/AddDeviceScreen/UITableViewControllerAddDeviceProperties.cs
