+++ Hestia/Hestia/Frontenda/DevicesScreen/EditDevice/UIViewControllerEditDeviceName.cs
