+++ Hestia/Hestia/Frontenda/DevicesScreen/UITableViewControllerDevicesMain.cs
