+++ Hestia/Hestia/Frontenda/EntryScreen/UIViewControllerLocalGlobal.cs
