+++ Hestia/Hestia/Frontenda/Local/UITableViewControllerServerDiscovery.cs
