+++ Hestia/Hestia/Frontenda/SettingsScreen/UITableViewControllerGlobalSettingsScreen.cs
